+++ CHANGELOG.md
@@ -16,3 +16,13 @@
 ## v0.2.4: Fixed generateSimpleEOFFiles. ##
 ## v0.2.5: Added initializeTemperature for generating initial temperature gradient ##
 * initializeTemperature
+## v0.3: Added chtMultiRegionFoamEpot and chtMultiRegionFoamEpotTransient solvers for multi region simulations with harmonic Elmer solver. ##
+* chtMultiRegionFoamEpot
+* chtMultiRegionFoamEpotTransient
+## v0.4: Removed duplicates 
+* buoyantFoamMHD
+* pimpleFoamMHD
+* simpleFoamMHD
+## Removed EOF file generation tools
+* generateSimpleEOFFiles
+* generatePimpleEOFFiles